chore: rename folder name change for now (to avoid merge conflict with nick PR)
